client/requirements.txt
jinja2
requests
PyYAML
plotly